+++ site/includes/addPage.inc.php
@@ -5,21 +5,17 @@
   */
   /**
   * This is a function for the class U.
-  * This function deletes a content
+  * This function adds a content
   * @see U For more informations about U.
   * @version Pb2.3Bfx0
   * @since Pb2.3Bfx0
   * @param string $table The name of the table where the page is stored. Default values: "Sites", "Blog"
-  * @param string $name Name of the page
+  * @param string $name The name of the content
   * @return bool True if succeeded, false if not
   */
-  function deletePage(string $table, string $name){
+  function addPage(string $table, string $name, int $authorID, date $date, number $online){
     global $U, $USOC;
-    // Checks if the page is index
-    if($table=="Sites" && $name=="index"){
-      return false;
-    }
-    $sql = "DELETE FROM ".$table." WHERE Name='".$name."'";
-    return mysqli_query($U->db_link, $sql);
+    // Checks if the name already exists
+
   }
 ?>